Almost got glm working. some tests are failing

+++ DirectionalLight.cpp
@@ -1,11 +1,11 @@
 #include "DirectionalLight.h"
 
 DirectionalLight::DirectionalLight()
-: Light("directional light", glm::vec3(0,0,0),Color(0.0,0.0,0.0),glm::vec3(0,0,0),true)
+: Light("directional light", glm::dvec3(0,0,0),Color(0.0,0.0,0.0),glm::dvec3(0,0,0),true)
 {
 }
 
-DirectionalLight::DirectionalLight(glm::vec3 pos, Color c,glm::vec3 dir)
+DirectionalLight::DirectionalLight(glm::dvec3 pos, Color c,glm::dvec3 dir)
 : Light("directional light", pos, c, dir,true)
 {
 }
