+++ Intersection.h
@@ -7,20 +7,20 @@
 class Intersection {
 public:
 	Intersection();
-	Intersection(glm::vec3 pos, Material m, double t, glm::vec3 n);
-	Intersection(glm::vec3 pos, Material m, double t, glm::vec3 n, int p_id);
-	void SetNormal(glm::vec3 n);
+	Intersection(glm::dvec3 pos, Material m, double t, glm::dvec3 n);
+	Intersection(glm::dvec3 pos, Material m, double t, glm::dvec3 n, int p_id);
+	void SetNormal(glm::dvec3 n);
 	virtual ~Intersection();
 	Material GetMaterial(){return m_material;}
-	glm::vec3 GetNormal(){return normal;}
+	glm::dvec3 GetNormal(){return normal;}
 	float GetT(){return m_t;}
-	glm::vec3 GetPosition(){return position;}
+	glm::dvec3 GetPosition(){return position;}
 	int GetId(){return m_p_id;}
 private:
 	Material m_material;
 	double m_t;
-	glm::vec3 position;
-	glm::vec3 normal;
+	glm::dvec3 position;
+	glm::dvec3 normal;
 	int m_p_id;
 };
