+++ PointLight.h
@@ -8,8 +8,8 @@ private:
 	double m_kc,m_kl,m_kq;//constant, linear, quadratic
 public:
 	PointLight();
-	PointLight(glm::vec3 pos, Color c, glm::vec3 dir);
-	PointLight(glm::vec3 pos, Color c, glm::vec3 dir,double kc, double kl, double kq);
+	PointLight(glm::dvec3 pos, Color c, glm::dvec3 dir);
+	PointLight(glm::dvec3 pos, Color c, glm::dvec3 dir,double kc, double kl, double kq);
 	virtual ~PointLight();
 	virtual bool IsPointLight(){return true;}
 	virtual bool IsDirLight(){return false;}
@@ -17,7 +17,7 @@ public:
 	void SetKc(double kc);
 	void SetKl(double kl);
 	void SetKq(double kq);
-	glm::vec3 GetAttenuation();
+	glm::dvec3 GetAttenuation();
 	double GetKc(){return m_kc;}
 	double GetKl(){return m_kl;}
 	double GetKq(){return m_kq;}
