+++ Triangle.h
@@ -5,20 +5,20 @@
 class Triangle : public Primitive {
 public:
 	Triangle();
-	Triangle(glm::vec3 verts[3], glm::vec3 norm, Material m);
-	Triangle(glm::vec3 verts[3], glm::vec3 norms[3], Material m);
+	Triangle(glm::dvec3 verts[3], glm::dvec3 norm, Material m);
+	Triangle(glm::dvec3 verts[3], glm::dvec3 norms[3], Material m);
 
 	virtual ~Triangle();
 	void print();
 	bool Intersects(Ray *r, Intersection *i);
 	bool IntersectsP(Ray *r, Intersection *i);
-	void printVector(glm::vec3 v);
-	void setFaceNormal(glm::vec3 n);
+	void printVector(glm::dvec3 v);
+	void setFaceNormal(glm::dvec3 n);
 
 private:
-	glm::vec3 vertices[3];
-	glm::vec3 normals[3];
-	glm::vec3 face_normal;
+	glm::dvec3 vertices[3];
+	glm::dvec3 normals[3];
+	glm::dvec3 face_normal;
 };
 
 #endif /* TRIANGLE_H_ */
